+++ tests/TestCase.php
@@ -5,11 +5,15 @@ declare(strict_types=1);
 namespace Hekal\ShipBridge\Smsa\Tests;
 
 use Hekal\ShipBridge\ShipBridgeServiceProvider;
+use Hekal\ShipBridge\Smsa\Contracts\SmsaGateway;
 use Hekal\ShipBridge\Smsa\SmsaServiceProvider;
+use Hekal\ShipBridge\Smsa\Testing\FakeSmsaGateway;
 use Orchestra\Testbench\TestCase as Orchestra;
 
 abstract class TestCase extends Orchestra
 {
+    protected FakeSmsaGateway $gateway;
+
     protected function getPackageProviders($app): array
     {
         return [
@@ -20,8 +24,12 @@ abstract class TestCase extends Orchestra
 
     protected function defineEnvironment($app): void
     {
+        $this->gateway = new FakeSmsaGateway;
+
+        $app->instance(SmsaGateway::class, $this->gateway);
+
         $app['config']->set('shipbridge.default', 'smsa');
-        $app['config']->set('shipbridge.drivers.smsa.base_url', 'https://smsa.test/v1');
-        $app['config']->set('shipbridge.drivers.smsa.token', 'test-token');
+        $app['config']->set('shipbridge.drivers.smsa.passkey', 'Testing0');
+        $app['config']->set('shipbridge.drivers.smsa.fake', true);
     }
 }